requirements_dev.txt
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
dacite==1.6.0
debugpy==1.5.0
decorator==5.1.0
docker==5.0.3
entrypoints==0.3
flake8==3.9.2
future==0.18.2
idna==3.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
jsonpath==0.82
jsonschema==4.0.1
jupyter-client==7.0.5
jupyter-core==4.8.1
lxml==4.6.3
matplotlib-inline==0.1.3
mccabe==0.6.1
nest-asyncio==1.5.1
networkx==2.6.3
nipyapi==0.16.3
numpy==1.21.2
packaging==21.0
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
pyaml==21.10.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==227
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
ruamel.yaml==0.16.12
six==1.16.0
tornado==6.1
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
websocket-client==1.2.1
xmltodict==0.12.0
